backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6


# Database
supabase>=2.25.1
psycopg2-binary==2.9.10

# Environment and Config
python-dotenv
celery
redis==1.0.0
pydantic>=2.11.7,<3.0
pydantic-settings>=2.3.0


# AI Models
openai==1.3.0
google-generativeai==0.3.0
groq==0.3.0
tiktoken==0.5.2

# Voice Services
elevenlabs==0.2.27
deepgram-sdk==3.2.7

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pypandoc==1.12

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and Monitoring
structlog==23.2.0

# HTTP Clients
requests==2.32.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx<0.28  # For async TestClient (compatible with FastAPI 0.104.1)


